LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/iframe-src-expected.txt



FAIL Replace before load, triggered by setting iframeElement.src assert_equals: onload must not yet have fired expected (undefined) undefined but got (boolean) true
